+++ src/main/java/com/management/library/controller/admin/AdminNewBookRequestController.java
@@ -1,7 +1,7 @@
 package com.management.library.controller.admin;
 
 import com.management.library.controller.request.newbook.dto.NewBookResultControllerDto;
-import com.management.library.dto.RequestSearchCond;
+import com.management.library.controller.dto.RequestSearchCond;
 import com.management.library.service.query.NewBookTotalResponseService;
 import com.management.library.service.query.dto.NewBookTotalResponseDto;
 import com.management.library.service.request.newbook.NewBookService;
