+++ src/main/java/com/management/library/repository/newbook/NewBookRequestRepositoryCustom.java
@@ -2,7 +2,7 @@ package com.management.library.repository.newbook;
 
 import static com.management.library.service.request.newbook.dto.NewBookRequestServiceDto.Response;
 
-import com.management.library.dto.RequestSearchCond;
+import com.management.library.controller.dto.RequestSearchCond;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
