+++ src/main/java/com/management/library/repository/rental/BookRentalRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.management.library.repository.rental;
 
 import com.management.library.domain.rental.Rental;
-import com.management.library.dto.BookRentalSearchCond;
+import com.management.library.controller.dto.BookRentalSearchCond;
 import com.management.library.service.rental.dto.RentalServiceResponseDto;
 import java.util.List;
 import java.util.Optional;
