+++ src/main/java/com/management/library/service/rental/RentalService.java
@@ -20,7 +20,7 @@ import com.management.library.domain.member.Member;
 import com.management.library.domain.rental.Rental;
 import com.management.library.domain.type.BookStatus;
 import com.management.library.domain.type.RentalStatus;
-import com.management.library.dto.BookRentalSearchCond;
+import com.management.library.controller.dto.BookRentalSearchCond;
 import com.management.library.exception.NoSuchElementExistsException;
 import com.management.library.exception.RentalException;
 import com.management.library.repository.book.BookRepository;
