+++ src/test/java/com/management/library/service/rental/RentalServiceTest.java
@@ -20,7 +20,7 @@ import com.management.library.domain.book.Book;
 import com.management.library.domain.rental.Rental;
 import com.management.library.domain.type.BookStatus;
 import com.management.library.domain.type.ExtendStatus;
-import com.management.library.dto.BookRentalSearchCond;
+import com.management.library.controller.dto.BookRentalSearchCond;
 import com.management.library.exception.RentalException;
 import com.management.library.repository.book.BookRepository;
 import com.management.library.repository.member.MemberRepository;
